backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Settings
pydantic-settings==2.6.1
python-multipart==0.0.12

# AI / Computer Vision
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
face-recognition==1.3.0
# insightface==0.7.3   # Uncomment for production (requires cmake)

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
